+++ tests/assets/extensions2.ts
@@ -1,0 +1,7 @@
+import type {} from 'sample-piral';
+
+declare module 'sample-piral' {
+  interface PiletApi {
+    registerFoo(): number;
+  }
+}